requirements.txt
streamlit
requests
fpdf2
python-docx
ollama
python-dotenv
